Move application raw mode test to separate file
